refactor(http): use shared code from public 'angular2/angular2' namespace
