+++ modules/angular2/src/http/base_response_options.ts
@@ -1,4 +1,4 @@
-import {Injectable} from 'angular2/src/core/di';
+import {Injectable} from 'angular2/angular2';
 import {isPresent, isJsObject} from 'angular2/src/core/facade/lang';
 import {Headers} from './headers';
 import {ResponseTypes} from './enums';
